Add files via upload
Code
Excel File containing the data from the rotating tool (must be present in the same location in order for the code to be able to read it)
Results, Matrix & Confusion Matrix from various models that I had tried to improve the performance. The third model is the one that I implemented as my final code.